wasm cleanup and extension

+++ cloudini_lib/include/cloudini_lib/wasm_functions.h
@@ -30,15 +30,15 @@
 extern "C" {
 
 // retrieves a YAML representation of Cloudini::EncodingInfo
-size_t cldn_GetHeaderAsYAML(uintptr_t encoded_data_ptr, size_t encoded_data_size, uintptr_t output_yaml_ptr);
+uint32_t cldn_GetHeaderAsYAML(uintptr_t encoded_data_ptr, uint32_t encoded_data_size, uintptr_t output_yaml_ptr);
 
 // Performs a full compression of the point cloud data, but return only the size of the
 // compressed data, not the data itself. Used mainly for testing purposes.
-WASM_EXPORT size_t cldn_ComputeCompressedSize(uintptr_t dds_msg_ptr, size_t dds_msg_size, float resolution);
+WASM_EXPORT uint32_t cldn_ComputeCompressedSize(uintptr_t dds_msg_ptr, uint32_t dds_msg_size, float resolution);
 
 // Preview the size of the decompressed point cloud data, needed to allocate memory in advance.
 // No actual decompression is performed.
-WASM_EXPORT size_t cldn_GetDecompressedSize(uintptr_t encoded_msg_ptr, size_t encoded_msg_size);
+WASM_EXPORT uint32_t cldn_GetDecompressedSize(uintptr_t encoded_msg_ptr, uint32_t encoded_msg_size);
 
 /**
  * @brief Given the a serialized DDS message containing "point_cloud_interfaces/CompressedPointCloud2",
@@ -50,8 +50,8 @@ WASM_EXPORT size_t cldn_GetDecompressedSize(uintptr_t encoded_msg_ptr, size_t en
  * @param output_data pointer for the output buffer where decompressed data will be written.
  * @return The size of the decompressed point cloud data, or 0 on failure.
  */
-WASM_EXPORT size_t
-cldn_DecodeCompressedMessage(uintptr_t compressed_msg_ptr, size_t msg_size, uintptr_t output_data_ptr);
+WASM_EXPORT uint32_t
+cldn_DecodeCompressedMessage(uintptr_t compressed_msg_ptr, uint32_t msg_size, uintptr_t output_data_ptr);
 
 /**
  * @brief Given compressed cloudini buffer, perform decoding.
@@ -63,8 +63,8 @@ cldn_DecodeCompressedMessage(uintptr_t compressed_msg_ptr, size_t msg_size, uint
  * @param output_data pointer for the output buffer where decompressed data will be written.
  * @return The size of the decompressed point cloud data, or 0 on failure.
  */
-WASM_EXPORT size_t
-cldn_DecodeCompressedData(uintptr_t encoded_data_ptr, size_t encoded_data_size, uintptr_t output_data);
+WASM_EXPORT uint32_t
+cldn_DecodeCompressedData(uintptr_t encoded_data_ptr, uint32_t encoded_data_size, uintptr_t output_data);
 
 /**
  * @brief Give a ROS2 DDS message containing "sensor_msgs/PointCloud2", perform compression
@@ -72,8 +72,8 @@ cldn_DecodeCompressedData(uintptr_t encoded_data_ptr, size_t encoded_data_size, 
  * The `output_data` is usually the "data" field of "point_cloud_interfaces::CompressedPointCloud2"
  * or can be saved directly to file.
  */
-WASM_EXPORT size_t cldn_EncodePointcloudMessage(
-    const uintptr_t pointcloud_msg_ptr, size_t msg_size, float resolution, uintptr_t output_data_ptr);
+WASM_EXPORT uint32_t cldn_EncodePointcloudMessage(
+    const uintptr_t pointcloud_msg_ptr, uint32_t msg_size, float resolution, uintptr_t output_data_ptr);
 
 /**
  * @brief Given the raw point cloud data, perform compression
@@ -81,6 +81,6 @@ WASM_EXPORT size_t cldn_EncodePointcloudMessage(
  * The `output_data` is usually the "data" field of "point_cloud_interfaces::CompressedPointCloud2"
  * or can be saved directly to file.
  */
-WASM_EXPORT size_t cldn_EncodePointcloudData(
-    const char* header_as_yaml, const uintptr_t pc_data_ptr, size_t pc_data_size, uintptr_t output_data_ptr);
+WASM_EXPORT uint32_t cldn_EncodePointcloudData(
+    const char* header_as_yaml, const uintptr_t pc_data_ptr, uint32_t pc_data_size, uintptr_t output_data_ptr);
 }
